@@ -22,13 +22,13 @@
     "mb-isochrone": "^0.1.0",
     "pluralize": "^8.0.0",
     "prop-types": "^15.7.2",
-    "react": "^16.12.0",
+    "react": "^18.2.0",
     "react-bootstrap": "1.0.0",
     "react-collapsible": "^2.6.2",
     "react-css-loaders": "^0.0.5",
     "react-datepicker": "^4.7.0",
     "react-debounce-render": "^6.0.0",
-    "react-dom": "^16.12.0",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^1.2.5",
     "react-fast-compare": "^2.0.4",
     "react-flip-toolkit": "^7.0.13",
