@@ -292,11 +292,10 @@ describe('ReportManager - ReportDetailView', () => {
   test('triggers onAddReport', async () => {
     const onAddReport = jest.fn();
 
-    AddReportMock = ({ onAddReport }) => {
+    AddReportMock = ({ onAddReport }) => { // eslint-disable-line
       useEffect(() => {
         onAddReport();
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-      }, []);
+      }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
       return null;
     };
@@ -327,7 +326,7 @@ describe('ReportManager - ReportDetailView', () => {
     executeSaveActionsMock = jest.fn(() => Promise.resolve(initialReport));
     executeSaveActions.mockImplementation(executeSaveActionsMock);
 
-    AddReportMock = ({ formProps }) => {
+    AddReportMock = ({ formProps }) => { // eslint-disable-line
       useEffect(() => {
         formProps.onSaveSuccess(addedReport);
       // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -370,7 +369,7 @@ describe('ReportManager - ReportDetailView', () => {
     executeSaveActionsMock = jest.fn(() => Promise.resolve(initialReport));
     executeSaveActions.mockImplementation(executeSaveActionsMock);
 
-    AddReportMock = ({ formProps }) => {
+    AddReportMock = ({ formProps }) => { // eslint-disable-line
       useEffect(() => {
         formProps.onSaveSuccess(addedReport);
       // eslint-disable-next-line react-hooks/exhaustive-deps
